Remove column factory
